# server/static/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Restaurants REST API",
      "description": "This is Flask Pizza Restaurants REST API for a Pizza Restaurant.",
      "license": {
        "name": "GNU 3.0",
        "url": "example.com"
      },
      "version": "1.0"
    },
    "externalDocs": {
      "description": "Find out more about Swagger",
      "url": "http://swagger.io"
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "tags": [
      {
        "name": "restaurants",
        "description": "See all restaurants and restaurant details"
      },
      {
        "name": "pizzas",
        "description": "Show all pizzas"
      },
      {
        "name": "restaurantpizzas",
        "description": "Add restaurant pizza details"
      }
    ],
    "paths": {
      "/restaurants": {
        "get": {
          "tags": [
            "restaurants"
          ],
          "summary": "Get list of all restaurants",
          "description": "Sends back JSON response with a list of all restaurants",
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Restaurant"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/restaurants/{restaurantID}": {
        "get": {
          "tags": [
            "restaurants"
          ],
          "summary": "Get a single restaurant",
          "description": "Returns a single restaurant",
          "parameters": [
            {
              "name": "restaurantID",
              "in": "path",
              "description": "ID of restaurant to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Restaurant"
                  }
                }
              }
            },
            "404": {
              "description": "Restaurant not found"
            }
          }
        },
        "delete": {
          "tags": [
            "restaurants"
          ],
          "summary": "Deletes a single restaurant",
          "description": "delete a single restaurant",
          "parameters": [
            {
              "name": "restaurantID",
              "in": "path",
              "description": "Restaurant id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RestaurantDeleted"
                  }
                }
              }
            }
          }
        }
      },
      "/pizzas": {
        "get": {
          "tags": [
            "pizzas"
          ],
          "summary": "Get list of all pizzas",
          "description": "Sends back JSON response with a list of all pizzas",
          "operationId": "Pizzas",
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Pizza"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/restaurantpizzas": {
        "post": {
          "tags": [
            "restaurantpizzas"
          ],
          "summary": "Creates a new RestaurantPizza entry",
          "description": "Creates a new RestaurantPizza that is associated with an existing Pizza and Restaurant",
          "operationId": "placeOrder",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestaurantPizza"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Pizza"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Restaurant": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "name": {
              "type": "string"
            },
            "address": {
              "type": "string"
            }
          }
        },
        "RestaurantDeleted": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string"
            }
          }
        },
        "Pizza": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "name": {
              "type": "string"
            },
            "ingredients": {
              "type": "string"
            }
          }
        },
        "RestaurantPizza": {
          "type": "object",
          "properties": {
            "price": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "pizza_id": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            },
            "restaurant_id": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        }
      }
    }
  }